Criacao das Classes CategoriaEmpresa, Endereco, Evento, Permissao, PontoTuristico, Reserva

+++ src/java/br/com/projeto/dao/EmpresaDAO.java
@@ -13,7 +13,7 @@ import java.util.List;
  */
 public interface EmpresaDAO {
     void save(Empresa empresa);
-    Empresa getEmpresa(Long id);
+    Empresa getEmpresa(int id);
     List<Empresa> list();
     void remove(Empresa empresa);
     void update(Empresa empresa);
